Error 'column reference ... is ambiguous' due to OUTPUT target list
When genrating output list for OUTPUT clause, same column can appear multiple times
in the list. This happens as different functions/expressions can reference same columns.
This leads to column reference ambiguity error for column aliases. As a solution, we do
not generate column alias if the underlying column is already part of the output list.

Task: BABEL-2522
Author: Avantika Dasgupta <davantik@amazon.com>
Signed-off-by: Harsh Lunagariya <lunharsh@amazon.com>